requirements.txt
1inch.py==1.9
aiofiles==23.1.0
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
base58==2.1.1
bitarray==2.7.3
certifi==2023.5.7
cffi==1.15.0
chardet==4.0.0
charset-normalizer==3.1.0
colored==1.4.4
cryptography==37.0.2
cytoolz==0.12.1
ecdsa==0.18.0
eth-abi==4.1.0
eth-account==0.9.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-retry==0.1.18
eth-rlp==0.3.0
eth-typing==3.4.0
eth-utils==2.1.1
exceptiongroup==1.1.1
frozenlist==1.3.3
gmqtt==0.6.12
h11==0.14.0
hexbytes==0.3.0
idna==2.10
ipfshttpclient==0.8.0a2
jsonschema==4.17.3
lru-dict==1.1.8
magic-filter==1.0.9
multiaddr==0.0.9
multicall==0.7.3
multidict==6.0.4
netaddr==0.8.0
outcome==1.2.0
parsimonious==0.9.0
protobuf==4.23.3
pycparser==2.21
pycryptodome==3.17
pyrsistent==0.19.3
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3
regex==2023.5.5
requests==2.30.0
rlp==3.0.0
secure-web3==1.3.2
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tblib==1.7.0
toolz==0.12.0
tqdm==4.65.0
trio==0.22.0
trio-parallel==1.2.0
trio-websocket==0.10.2
typing_extensions==4.5.0
urllib3==1.26.15
uvloop==0.17.0
varint==1.0.2
web3==6.4.0
websockets==11.0.3
wsproto==1.2.0
yarl==1.9.2
